requirement.txt
APScheduler==3.7.0
asgiref==3.3.4
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-mysql==3.11.1
future==0.18.2
idna==2.10
line-bot-sdk==1.19.0
Pillow==8.2.0
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tzlocal==2.1
urllib3==1.26.5
